Homework 37.7

+++ app/src/main/java/ru/vodolatskii/movies/data/entity/Movie.kt
@@ -15,6 +15,6 @@ data class Movie(
     val name: String = "",
     val description: String = "",
     val posterUrl: String = "",
-    val rating: Float = 0.0f,
+    val rating: Double = 0.0,
     var isFavorite: Boolean = false
 ) : Parcelable
